requirements.txt
SQLAlchemy
python-dotenv
fastapi
python-jose
pydantic
gunicorn
pymysql
pydantic[email]
httpx
python-multipart
pydantic-settings